Request attributes from controller upon failure to obtain lock
It is possible that when multiple targets are created / try to trigger distribution set assignment checks at similar times, there is a deadlock. When this happens, a target continues to stay in REGISTERED until the next time it sends attributes. This fix will request attributes from target upon deadlock so that the delay to auto assignment is minimized.